requirements.txt
opencv-python==4.10.0.84
numpy>=1.24,<3
PyYAML>=6.0
Flask>=2.0.0
flask-cors>=3.0.0
Pillow>=9.0.0
rembg>=2.0.0
Werkzeug>=2.0.0
numpy>=1.21.0
onnxruntime>=1.15.0